requirements.txt
# Hadron Core Sovereign Dependencies v9.0.0

# CORE FRAMEWORK
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# AI & NEURAL REASONING
google-genai>=1.0.0
tenacity>=8.3.0

# METABOLIC MONITORING
psutil>=5.9.0
prometheus-client>=0.20.0

# TELEMETRY & LOGGING
structlog>=24.1.0

# UTILITIES
typing-extensions>=4.10.0
aiohttp>=3.9.0
cryptography>=42.0.0
